[feat]: Add CustomForm and refactor forms in pages
