+++ frontend/src/components/NavigationLinks.js
@@ -4,8 +4,8 @@ import {Link} from "@chakra-ui/react";
 
 const pages = [
     {href: "/", text: "Home"},
-    {href: "/signup", text: "Sign Up", testId: "SignUpLink"},
     {href: "/remittance", text: "Transfer", testId: "RemittanceLink"},
+    {href: "/signup", text: "Sign Up", testId: "SignUpLink"},
     // Add more links as needed
 ];
