+++ frontend/src/index.css
@@ -5,10 +5,3 @@
     border-radius: 8px;
     box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
 }
-
-.page form {
-    display: flex;
-    flex-direction: column;
-}
-
-
