+++ frontend/src/utils/AppRouter.js
@@ -12,8 +12,8 @@ import Loading from '../components/Loading';
 //Pages
 const Home = lazy(() => import('../pages/Home'));
 const Login = lazy(() => import('../pages/Login'));
-const RemittancePage = lazy(() => import('../pages/RemittancePage'));
-const CreateUser = lazy(() => import('../pages/CreateUser'));
+const RemittancePage = lazy(() => import('../pages/Remittance'));
+const CreateUser = lazy(() => import('../pages/SignUp'));
 const Kyc = lazy(() => import('../pages/KYC'));
 const DashBoard = lazy(() => import('../pages/DashboardPage'));
 const NotFound404 = lazy(() => import('../pages/ErrorPage'));
